[libromdata] Wim: Fix two crashes on fuzzed data.
Crash #1 was caused by using the header's image count instead of the
actual vector size.

A corrupted (or maliciously crafted) WIM XML file could have an
incorrect image count field. This crashes in the second half of the XML
parsing when processing the vector, since the index goes out of bounds.

Fix this by using a ranged-based `for` loop using the vector, and
maintain an index count separately.

Crash #2 was caused by not limiting the image count to a reasonable
maximum. A fuzzed value of 2147483650 caused an std::bad_alloc when
calling std::vector<>::reserve().

Other changes:
- Use a constant for the maximum image size.
- Get rid of the rowloop macros.

Found using aflplusplus-4.07c.