[libromdata] KeyStoreUI::importKeysFromBin(): Remove an unnecessary UNREF().
We're already using unique_RefBase<> here, so a manual UNREF() isn't
needed. In addition, this doesn't NULL out the pointer in the
unique_RefBase<>, so it could actually result in a double-free.
(Probably won't happen, since the UI frontend should have already
verified that the file exists, since it used an OS-native file picker
dialog...)