[libromdata] loadInternalImage(): Fix up some return values and *pImage clearing.
In a few code paths, 0 was returned if an error occurred instead of
a more appropriate error code, e.g. -ENOENT.

In other code paths, *pImage was not set to nullptr on error.

This *could* have caused an assertion in RomData::image(), but I haven't
seen any assertions, probably because the errors are filtered out in
e.g. rpcli.

Affected classes:
- DreamcastSave
- GameCubeSave
- WiiSave
- WiiWIBN
- iQuePlayer
- Nintendo3DS
- NintendoBadge