[kde] ITab.hpp: Remove NOVTABLE.
I didn't notice any issues in local builds, but on AppVeyor, the KDE4
build is failing because common.h isn't included at this point, so it
can't figure out what NOVTABLE is.

NOVTABLE is currently only useful in MSVC builds, so just remove it from
KDE ITab.hpp for now.