[libromdata] SPC: Get the total duration for both ID666 and extended ID666.
For ID666, the loop field is used, and loop count is set to -1.
If loop count is -1 and the other values aren't set, then the
loop duration is used as the total duration, plus fadeout.
Otherwise, the loop duration is multipled by loop count and added
to the intro+end, plus fadeout.

TODO:
- Binary ID666 has not been tested.
- Metadata.