[librptexture] SegaPVR: Initial implementation of GameCube CI4.
This is from a very old stash in my local git repository, modified
to work correctly with all of the changes since then.
[2017/07/16 02:21 PM EDT]

This has an external color lookup table, so we'll use a grayscale
RGB5A3 palette for now.

FIXME: Something's off with the tiling. Puyo Tools (git master)
handles it correctly.