[libromdata/tests] ImageDecoderTest: Add mipmap tests, starting with KTX2.
If a mipmap level is specified, IMG_INT_IMAGE is always used.

If the mipmap level is 0, then the rp_image is compared to the "regular"
rp_image. The rp_image objects for both mipmap level 0 and the "regular"
image must be the same.

Add PNG reference images for KTX2 mipmap test images. The original
reference image has been renamed as mipmap level 0.