[librptexture] PowerVR3: Use a temporary variable to adjust for mipmap levels.
If we use `mip`, we end up always storing the image in mipmaps[0].

Prior to the use of shared_ptr<>, this could have caused a memory
leak, since it would overwrite whatever was in mipmaps[0] without
unreferencing it, but nothing used any mipmaps other than 0 before
the addition of rpcli's `-m` option.

Fixing this is needed in order to fix the upcoming PowerVR3 mipmap tests
in ImageDecoderTest, which verify that retrieving the same mipmap twice
results in getting the same rp_image object.