[libromdata] EXEData: Add 0x0601 for PowerPC (big-endian).
This CPU type is used by the old MSVC for Mac.

0x01F2, the CPU type for Xbox 360, is now "PowerPC (big-endian; Xenon)".

FIXME: Subsystem version on the test DLL is "1024.32768". Byteswapping
this results in "4.128", which is still not quite right. Import and
export tables aren't showing up either, which probably means some
fields are in big-endian that are normally in little-endian.

Fixes #396: PE machine value 0x0601 == PowerPC big-endian (classic Mac)
Reported by @Wack0.