[gtk] FlagSpriteSheet: New ISpriteSheet for flags.
LanguageComboBox: Rework to use FlagSpriteSheet.

NOTE: The code paths for "sprite sheet could not be loaded" have been
removed. This shouldn't happen, and it just overcomplicates things.
If the sprite sheet can't be loaded, the loaded icons will be nullptr,
so that code path will be taken instead.

LanguageComboBox's flags now show up properly on GTK4 with GdkTexture.
The GTK3 and GTK2 versions still work, too.