[gtk] CreateThumbnail: Inline the pixel access instead of using PIMGTYPE_get_image_data() and PIMGTYPE_get_rowstride().
GdkTexture doesn't allow direct access, so we have to download the
pixel data using gdk_texture_download().

FIXME: Scaling isn't implemented, so we're getting a "thumbnail" that's
actually just the top left of the image.

FIXME 2: Use GdkTextureDownloader to ensure it's not premultiplied and
to swap the channels. We're getting an R/B swapped image right now.