[gtk] PIMGTYPE: Un-inline PIMGTYPE_scale().
This will be implemented later. I initially un-inlined it in the previous
commit while debugging stuff to see if duplicating the texture instead of
simply referencing it would fix things. It turns out the issue was that
I was g_free()'ing `pixels`, which is adjusted by the row_pointers code.
Using a separate `texdownload` variable that always contains the base
address of the texture and then freeing *that* fixed it.

The un-inlined version will eventually be used once I implement scaling,
probably by mixing in the GTK3 Cairo code.