[win32] Port ISpriteSheet from the GTK version.
This helps to clean up a lot of code, especially since the Win32 version
is quite a bit more "manual" than the GTK version.

Note that getFilename() has been changed to getResourceID(), since Win32
resources use an ID scheme instead of a filename scheme. Also, getIcon()
takes an optional DPI value to encode in the HBITMAP. Not sure if that
actually does anything, but we were doing it before, so keep doing it.