.github/workflows/codecov.yml: Initial Codecov.io workflow.
Building in the project root directory so Codecov.io finds the
coverage files. Not sure if it'd work with a separate build/
directory...

Note that the `lcov` package is needed for rom-properties' built-in
coverage target. It technically isn't needed for Codecov.io, but
the build scripts will fail if it's missing.

Root codecov.yml: Ignore extlib/.

NOTE: libromdata is currently built as a shared library in coverage
builds, so LD_LIBRARY_PATH is set in .github/workflows/codecov.yml .