[libromdata/tests] RomHeaderTest: Add 3DSident v0.8.0 .3dsx and .cia files.
For initial Nintendo3DS code coverage.

Note that the .cia files are not encrypted.

Code coverage changes:

Before:
  lines......: 24.0% (12795 of 53301 lines)
  functions..: 21.9% (784 of 3586 functions)

After:
  lines......: 25.7% (13689 of 53301 lines)
  functions..: 23.7% (851 of 3586 functions)

NOTE: The full .3dsx and .cia files are included for now. If any more
files are added, I'll need to figure out a decent way to zero out
portions of the files that aren't needed for testing purposes.