[libromdata/tests] ImageDecoderTest: Add the remaining Blender DDS images, with rpcli-created PNG reference images.
The differences between both imagemagick and gimp (for images they
actually support) are usually an off-by-one on one or more channels.
This is due to floating-point differences in DXTn decoders, and I'm
not sure if there's anything that can be done about it, aside from
maybe rewriting my DXTn decoder to match others. (I'm not sure how
complicated or necessary this is, so I'm leaving it for now.)

Code coverage changes:

Before: (commit 676ca94e18e005732585899db373281245c6b750)
  lines......: 25.8% (13764 of 53308 lines)
  functions..: 23.8% (855 of 3586 functions)

After:
  lines......: 25.9% (13809 of 53318 lines)
  functions..: 23.9% (856 of 3586 functions)

TODO: Test mipmaps and also the "broken" files.