[libromdata/tests] RomHeaderTest: Add several DDS images.
Removed most of the large, uncompressible DDS images, thouhg this should
still have a good amount of test coverage.

The differences between both imagemagick and gimp (for images they
actually support) are usually an off-by-one on one or more channels.
This is due to floating-point differences in DXTn decoders, and I'm
not sure if there's anything that can be done about it, aside from
maybe rewriting my DXTn decoder to match others. (I'm not sure how
complicated or necessary this is, so I'm leaving it for now.)

Code coverage changes:

Before:
  lines......: 25.9% (13809 of 53318 lines)
  functions..: 23.9% (856 of 3586 functions)

After:
  lines......: 26.0% (13869 of 53318 lines)
  functions..: 24.1% (863 of 3586 functions)