[libromdata/tests] ImageDecoderTest: Get pixel format and DX10 format directly from RpTextureWrapper.
[libromdata] RpTextureWrapper: Add accessor functions for pixel format
and DX10 format. This does result in additional exports that are only
used for test suites, but it reduces overhead by getting the pixel format
directly and not having to search through RomFields.

FIXME: DX10 format still requires searching through RomFields. Fixing this
would require adding a function to FileFormat just for DX10 format.

I made this change because I noticed the code coverage barely changed with
the addition of DDS files to RomHeaderTest. It turns out this was because
ImageDecoderTest was retrieving fields for DirectDrawSurface (and other
texture formats) in order to verify the pixel format and, for DX10 DDS
files, the DX10 format. RpTextureWrapper now exports two functions to do
this instead, and pixelFormat() doesn't require getting RomFields.

Code coverage changes:

Before:
  lines......: 26.0% (13869 of 53318 lines)
  functions..: 24.1% (863 of 3586 functions)

After:
  lines......: 25.4% (13541 of 53331 lines)
  functions..: 23.8% (854 of 3588 functions)

This *reduces* the code coverage by 328 lines and 9 functions, but those
lines weren't actually being tested to ensure they were valid. Need to add
more RomHeaderTest cases to verify that it's working correctly.