[win32] KeyManagerTab: Use dark mode for the ListView and Import button.
For the ListView, we need to get the alternate row color using the
ListView handle as a reference instead of GetSysColor(COLOR_WINDOW).
Otherwise, it will always get the "light" version.

RP_ShellPropSheetExt: Don't cache colorAltRow for now, since we
might have multiple ListViews.
- TODO: Figure out a better way to cache it?

LibWin32UI: Renamed getAltRowColor(0 to ListView_GetBkColor_AltRow().
It takes an HWND to a ListView as a parameter.

TODO:
- WM_THEMECHANGED handling everywhere.
- Backport just this to master, even if we don't merge the whole
  branch due to an incomplete Win32 dark mode implementation?