[win32] KeyManagerTabPrivate::ListView_CustomDraw(): Fix the row background color when an icon is showing.
This doesn't affect Windows 7 or Windows XP, but it definitely affects
Windows 10 in both light and dark modes.

Don't erase the background color unless the row is neither selected
nor highlighted and is an odd-numbered (alternate) row. This allows
us to keep the original background theme. All other methods I tried,
including GetThemeColor() and DrawThemeBackgroundEx(), didn't work
correctly.

Note that we have to check if the "hot" item matches dwItemSpec,
since uItemState is 0 if the user highlights a different column in
the same row, even though we have LVS_EX_FULLROWSELECT set.