[win32] AchievementsTab: Fix icon transparency on alternating rows.
We need to draw the row background in the CDDS_SUBITEM prepaint event.
Note that the rectangle for subitem 0 seems to leave a small border on
the left side, so we have to get the RECT manually.

Add only the original bitmap (without icon mask) to the ImageList.
The mask isn't needed, since it's 32-bit ARGB, and the actual problem
was the missing CDDS_SUBITEM prepaint event.

In testing, there doesn't seem to be any change on Windows 7 (ListView
icons are still not transparent), but it definitely works on Windows 10.
(Tested on 1809 and 21H2.)

TODO: Port this to RP_ShellPropSheetExt for RFT_LISTDATA.