[zlib-ng] Disable TLS when building for Windows on i386. Windows XP has problems with it.
rp-config suddenly started crashing on XP when trying to decode PNG
images. It turns out that *something* caused zlib-ng to start using
TLS, though the code indicates it may have always used it...

It's entirely possible that this was broken on XP for quite a while,
and I just didn't notice it until now.

TLS usage was added in zlib-ng 2.1.0-beta1. rom-properties updated
from v2.0.7 to v2.1.2 in rom-properties 2.2, so the previous release
will randomly crash on Windows XP.