[win32] RP_ShellPropSheetExt: Restore lvBgColor, but with 0xFF000000.
The alpha channel was previously set to 0, which is why it didn't work.

FIXME: Not sure if alpha-transpraent achievement icons work properly.
Attempting to port the CustomDraw stuff from AchievementsTab resulted
in subitems of odd rows disappearing on mouseover. I think this may
have happened in AchievementsTab too (in light mode)...