[kf5] Use the newer registerPlugin<> with no keywords on *all* versions of KF5.
The keyword version was deprecated in KF5 5.89, but it's been supported
since KF5 5.0. For some reason, on Ubuntu 18.04, XAttrView wasn't
showing up at all with the keyword version, even though the main
rom-properties tab did. Removing the keyword fixed it.

TODO: Check if KDE4 actually requires the keyword paramter. If it doesn't,
then we should remove it from the KDE4 version, too.