[kde] Ported the MS-DOS attribute writing functionality to the KDE version.
This one's a bit different compared to the GTK version because it
doesn't immediately write changes. Instead, if the user changes one
of the RHAS attributes, a "modified" signal is emitted, which sets
the property page as "dirty". When the user clicks "OK" in the dialog
(KF5 5.44; possibly "Apply" in newer versions?), applyChanges() is
called in the plugin.

In order to handle the case of multiple XAttrView instances being
loaded (not likely), an std::set is used to store every XAttrView.
- On applyChanges(), the set is iterated and XAttrView::applyChanges()
  is called on each one.
- On destroyed(), the XAttrView is removed from the set.

TODO: Windows version? These attributes are normally adjustable by the
user in the Windows file properties dialog, but we might as well add it
for completeness. Also, Windows Explorer doesn't allow setting the
"System" attribute. (TODO: Check if Win95 allows it; Win98 shows it but
doesn't allow modification, I think Win2000 got rid of the checkbox.)