[libromdata] GameCube, WiiWAD: Clarify that age ratings are Wii only, not GameCube.
GameCube: Add a new bool hasRvlRegionSetting that indicates if the region
setting field was read. This was previously overloaded with hasRegionCode,
which only indicates if a region code is available. (GCN has region codes,
but not age ratings.)

GameCube, WiiWAD: Fix comments to indicate that only Wii has age ratings.