[libromdata] GameCube: Convert uint34_rshift2_t into a struct with helper functions.
With the typedef, it was possible to accidentally manipulate it as if
it was a uint32_t. With the struct, it has to be converted to either
uint64_t or off64_t before it can be accessed.

This shouldn't have any functional changes.

wii_structs.h: Add struct uint34_rshift2_t with C++ helper functions:
- get(): Get as uint64_t.
- get_be(): Get as uint64_t, byteswapped from big-endian if necessary.
- geto(): Get as off64_t.
- geto_be(): Get as off64_t, byteswapped from big-endian if necessary.
- set(): Set from uint64_t.
- seto(): Set from off64_t.

NOTE: Only geto_be() is used right now.
TODO: Add set_be() and seto_be()?