[svrplus] svrplus.c: Revert the previous 32-bit ARM change; move the check to CPU_arm64.
If IsWow64Process2() returns 32-bit ARM, then that means it's running
on a 32-bit ARM system. There's no point in checking for Windows 11
because Windows 11 dropped support for 32-bit ARM CPUs.

In the CPU_arm64 path, check for Windows 11 build 25905, and if it's
found, don't register for 32-bit ARM.

See #398.