Remove BUILD_GTK2; use BUILD_XFCE everywhere instead.
This broke the GitHub CodeQL and CodeCov.io builds because we're not
installing GTK+ 2.x in those builds, and BUILD_GTK2 isn't a cached
variable; it's a "regular" variable.

The newly-added GTK tests expected BUILD_GTK2 to be unset if GTK2 wasn't
actually found, but this wasn't the case, so it failed to find the
GTK2::gtk target.

[res] Also check for BUILD_GTK4 in addition to BUILD_XFCE and BUILD_GTK3.
Not likely to have caused any problems yet, since no one has a "pure"
GTK4 system yet.