[amiibo-data] amiiboc.cpp: Use _tmain() and other Unicode functions on Windows.
NOTE: Only the statements that needed to be changed to Unicode have been
updated. Some more Unicode conversion might be done in another commit,
but since the amiibo-data.txt file is UTF-8, not all of it will be changed.

CMakeLists.txt: Remove an obsolete TODO about running amiiboc to compile
the amiibo database.