[librptext] specialized.cpp, str8_to_utf8(): Fix an off-by-one for zero-length strings.
If the first byte is 0x00, it incorrectly calculates a string length of 1.
This results in the first byte being converted. In the case of the
Commodore PETSCII character sets, 0x00 is mapped to U+FFFD, so it would
always show a Unicode replacement character for CBMDOS disks with an
empty disk name.

This wasn't noticed before because nothing other than TextFuncsTest
actually made use of the specialized code pages, and TextFuncsTest
only used CP_RP_ATARIST, which has 0x00 mapped to U+0000.

TODO: Add PETSCII tests and zero-length string tests.