[libromdata] CBMDOS: Add support for D80 and D82 images.
C8050 and C8250 are quad-density drives.
C8050 is single-sided; C8250 is double-sided.

Compared to C1541, the header sector is different (and doesn't have
the BAM), but the directory entries are the same. To compensate for
this, the header parser has been reworked to get the offsets from
the header sector based on disk type.