[libromdata] CBMDOS: Allow early missing tracks in G64 images.
Monopoly (USA) (C64).g64 is missing tracks 1-4. Previously, this would
cause it to simply stop reading *any* track information, which resulted
in an assertion due to going out of bounds in the track_offsets vector.

Added some extra checks to prevent the out-of-bounds condition from
happening.

FIXME: It's not loading the directory from this disk image...