[gtk,kde,win32] Allow empty RFT_LISTDATA fields.
In the debug build, an assertion was tripping when viewing a C64
disk image that has no files.

In the release build, this would have resulted in the directory field
simply not showing up.

Now, it shows up with an empty list.

For other RomData subclasses, an empty list is probably not right,
but it's better to have it show up than to not show up at all.