[libromdata] DpfReader: Initial reader class for GameCube DPF and Wii RPF files.
These are sparse disc images created by the official SDKs.

Note that unlike the other sparse disc image formats we support,
DPF and RPF both have byte granularity instead of block granularity.
Therefore, we're not using SparseDiscReader as a base class.

RomDataFactory: Use IDiscReader instead of SparseDiscReader as the base
class in order to be able to use DpfReader.

Added DPF and RPF MIME types and file extensions for the GameCube class.
Also added the MIME types to the xdg/ directory.

FIXME: memset() in read() is causing an Internal Compiler Error when
compiling with gcc-13.2.0. Using a (slower) `for` loop instead.

TODO:
- RPF images don't start at virt=0, phys=0. Need to add a fake block for
  phys=0 and adjust all the virtual offsets.
- Pre-adjust all phys offsets by adding the data start offset? Doing this
  might just waste cycles if we don't read that much data from the disc.