[libromdata] GameCubeBNR: Improve detection of Shift-JIS.
shouldHandleStringAsShiftJIS() now takes the first character of the
string instead of just a bool indicating if the first character is
'\xA9' (cp1252 copyright symbol).

If the region is USA/EUR, and the first character is a Shift-JIS
first byte that either isn't allocated or is rarely used in cp1252,
assume it's Shift-JIS.

This fixes game info display for rubysap-timerfix-1219-10-eng.dpf,
which has a game ID of "RELS01".