[libromdata] Initial ColecoVision parser.
ColecoVision ROMs have a short magic number, either AA 55 or 55 AA,
so the file extension is used for detection.

For ROMs that have AA 55, the ColecoVision logo is displayed on startup,
and the game name is displayed. The game name is stored at 0x0024,
and is formatted using three lines. The first two lines are displayed
in reverse order, and the last line is a 4-digit year.

TODO:
- Improve formatting.
- Add more fields.