[librpbase] gtest_init.cpp: Add arm_fadvise64_64() to the seccomp filter.
RomHeaderTest uses fadvise() to advise the system that we're reading
files sequentially. On 32-bit ARM, 64-bit values must be passed in a
specific way, and the standard fadvise64() syscall isn't suitable for
this. arm_fadvise64_64() is basically the same thing, but it rearranges
the parameters to fit the ARM EABI.

This should fix RomHeaderTest in 32-bit ARM builds.