[libromdata] PalmOSPrivate::loadIcon(): Don't byteswap the bitmapType width/height fields in the struct directly.
This is confusing, since none of the other fields are byteswapped
directly in the bitmapType struct.

Not doing the byteswapping might result in *slightly* more overhead,
but it's easier to understand what the code is doing.

Also, the initial size check was checking for `height >= 0` instead of
`height > 0` for some reason. (The assertion was correct, though...)
Fixed that so it correctly checks for `> 0`.