[libromdata] PalmOS: Factor out the compressed data size code from the decompress_*() functions.
loadIcon() now loads the compressed data size and makes use of it when
loading the compressed icon data.

This simplifies the code and allows us to load less data from the file
when it uses a compressed icon.

compr_data_len = compressed icon size (same as uncompressed if not compressed)
icon_data_len = uncompressed icon size