[kde] Split RomThumbnailCreator into a separate shared library. (KF5 5.100+; KF6)
This was initially done in an attempt to figure out why Dolphin is
ignoring the checkbox in the "Previews" configuration (it *always*
enables rom-properties, even if it's unchecked). I haven't figured
it out yet, but we might as well keep it separate, even if it means
there's a bit of redundant code.

This does not affect RomThumbCreator (the old version), which is
still in rom-properties-kde4 and rom-properties-kf5, nor does it
affect the entry point for rp-stub.