[libromdata] WiiUPackage: Very preliminary Wii U NUS package parser.
This is the first "directory" parser, so nothing's actually implemented
other than detecting the fact that the directory is a Wii U NUS package.
It does this by looking for title.tik, title.tmd, and title.cert.

RomDataFactory: If the specified filename is a directory, call
WiiUPackage::isDirSupported_static().

[librpfile] FileSystem.hpp: Export is_directory().
- TODO: Bump the SOVERSION?

[rpcli] rpcli.cpp: If it's a directory, call RomDataFactory::create()
with the filename.

Invoking `rpcli` on an NUS package now shows that it is an NUS package,
but it doesn't show any ROM fields yet.

TODO:
- Directories on Windows, KDE, and GTK.
- Option to disable directories because it might cause slowdown.