[libromdata] WiiUPackage: Add file loading and load the FST.
The FST is the "bootable" content, which is usually the first one.

New function openContentFile() to open the specified content file.
The content index here is the index in the TMD.
- TODO: H3-hashed files.

New function reset() to clear various variables. Used by the private
class constructor to reduce lines.

Private class destructor: Delete ticket and fst.

TODO: Decrypt the title key. Using a hard-coded title key for now.
(It's all zeroes in this commit; replace it with your own.)