[libromdata] WiiUPackage: Clear d->isValid if Ticket/TMD failed to load; keep going if decryption fails.
Even if decryption fails, we can still show Ticket and TMD information
later on.

Need to clear d->isValid if Ticket or TMD fail because otherwise, the
debug build will assert because d->fields is nullptr. Also, we don't
want the package to be "valid" if either Ticket or TMD are invalid.