[libromdata] WiiPartition: Use WiiTicket to decrypt the title key.
d->initDecryption() is now available for no-crypto builds, though only
the parts that check for an unencrypted disc image.

The title key is no longer cached in d->title_key.

WiiTicket::encKey() no longer requires calling decryptTitleKey(),
which allows it to function in no-crypto builds.

FIXME: Attempting to read an encrypted Wii disc image in the debug build
results in an assertion failure in GcnFst.