[libromdata/tests] SuperMagicDriveTest needs librpcpuid on i386.
The SMD decoder has implementations for MMX and SSE2. Both of these are
available by default on amd64, so librpcpuid isn't needed on amd64.
On i386, however, librpcpuid *is* needed.

Interestingly, this only broke the Windows 32-bit builds on AppVeyor.
The 32-bit builds may have succeeded due to the use of IFUNC, which
requires librpcpuid in libromdata but not in SuperMagicDriveTest.