[libromdata] WiiUPackage: Display age ratings.
The age ratings are stored in meta.xml as individual fields, instead of
a simple easy-to-parse binary block. Joy.

The values are slightly different compared to Wii and 3DS, but it uses
the same ordering, and even has the same deprecated (and unused) BBFC
and Finland/MEKU fields.

- 0x00-0x1F: Age rating
- 0x80: No rating (opposite of Wii and 3DS!)
- 0xC0: Rating pending

NOTE: "0" indicates no age restriction, but it can be confused with some
actual ratings that have a "0" value. (Not sure if Wii U has the same 0x20
"No Restriction" bit as 3DS...)

RomFields:
- AgeRatingsCountry: Add New Zealand (OFLC).
- ageRatingAbbrev(): Likewise.