[gtk] MessageWidget: Add wrappers for transition duration and type.
GtkRevealer defaults to GTK_REVEALER_TRANSITION_TYPE_SLIDE_DOWN.

KeyManagerTab: Explicitly set the transition type to DOWN.

RomDataView_ops.cpp: Explicitly set the transition type to UP. This fixes
the transition, since the MessageWidget is located at the *bottom* of the
window. Previously, it would shrink from the top; the label stayed where
it was. Now, the label moves along with the rest of it, as it should.