[gtk] rom-properties-glib: Set GLib version to 2.34, or 2.66 if GTK4 is enabled.
This *is* definitely what's causing the size difference. With 2.34, it
has the smaller XFCE/GTK3 size; with 2.66, it has the slightly larger
GTK4 size. Not entirely sure *what* specifically is increasing it...

Either way, building for glib-2.66 won't affect the XFCE or GTK3 plugins,
since it only affects internal data structures and not linkage.