[win32] RP_XAttrView: Call config->getBoolConfigOption(), not config->getBoolConfigOption_default().
We want to read the *current* setting, not the default setting.

Note that I forgot to add the checkbox to rp-config, but it will still
read the value from the configuration file if it's present.

This fixes a regression from commit eb54f1d727dcb7dddec5a59f5a18353309239593.
([librpbase] Config: Use a single function to get boolean configuration options (and their defaults).)