[win32] OptionsTab: Add checkboxes for ShowXAttrView and ThumbnailDirectoryPackages.
Rearranged some code to match the UI order.

Added the checkboxes to all of the localized *.rc files, though the
strings are not localized at the moment.

RP_ExtractIcon, RP_ExtractImage: If ThumbnailDirectoryPackages is
disabled, check if the file is a directory. If it is, don't thumbnail it.

NOTE: IThumbnailProvider and IPropertyStore take a stream, so they can't
thumbnail (or extract metadata from) directories.