[gtk] stub-export.c: Handle the "OK" button in the dialog.
FIXME: On GTK4, we're getting a warning on exit:

Gtk-WARNING **: 01:18:44.611: A floating object was finalized. This means that someone
called g_object_unref() on an object that had only a floating
reference; the initial floating reference is not owned by anyone
and must be removed with g_object_ref_sink().