[gtk] GTK4: Unreference the GtkSingleSelection object after setting it in GtkColumnView.
GtkColumnView takes a reference, but does not take ownership.
Hence, we have to unreference it to relinquish ownership.

This probably doesn't fix any memory leaks per se, since the only GTK4
stuff that's currently user-visible in normal operation is AchievementsTab,
and when that's closed, the program exits.

rp_rom_data_view_update_multi_RFT_LISTDATA_MULTI(): ref the selModel
before NULLing it in the GtkColumnView. Otherwise, it will end up
getting deleted due to losing all references. (Unref it after re-adding
it to the GtkColumnView.)