[gtk] RomDataView_gtk4.cpp: Enable sorting.
This uses GtkSortListModel.

A GtkCustomSorter is added to each GtkColumnView. GtkCustomSorter is a
GtkSorter implementation that takes a function pointer to a sorting
function. The sorting function, GCompareDataFunc, is similar to
GtkTreeIterCompareFunc, except it provides two items (in this case,
RpListDataItems) and user_data; no model pointers.

[gtk/tests] SortFuncsTest: TODO: An actual GTK4 version that tests
the GTK4 sorting functions.