[gtk4] Need to unref() items after appending them to GListStore.
Otherwise, the items might leak.

GtkListStore also required this, but it didn't require custom item
classes; it took strings directly. Hence, I didn't notice it as much.
For PIMGTYPE, I had to manually unref() the image after adding it
to GtkListStore, so I should've realized I needed to do this for
GListStore in GTK4. Meh.