[librpbase/tests] TextFuncsTest: Convert the test arrays to std::array<>.
Removed most casting and ARRAY_SIZE() usage. The 8-bit text data still
needs to be cast to char, since it's stored as uint8_t in order to allow
us to make use of the high bit.

NOTE: The Atari ST and ATASCII test data used strings instead of arrays.
Converted to arrays in order to be usable with std::array<>.

NOTE 2: The utf16_* arrays are actually uint8_t in order to maintain the
correct byte-order regardless of host platform. Hence, casting is still
needed here. This also requires use of C16_ARRAY_SIZE() to determine the
size in char16_t units.

NOTE 3: Some char strings in TextFuncsTest.cpp were left as-is.
I might convert them to std::array<> later.