[gtk] LanguageComboBox: Send a property notification for "selected-lc" instead of a signal.
Qt uses custom signals for property changes. GObject prefers using
notification signals, which are automatically created and registered
when a property is registered. Let's use a GObject signal.

GtkComboBox, which is deprecated in GTK3, sends a "changed" signal.
GtkDropDown, which is new in GTK4, sends "notify::selected" instead.

Note that the notify signal handler doesn't get the actual value that's
set, so we have to get it from the object. Not a big deal.