[libromdata] DMGSpecialCases: Remove all constexprs that were just added.
For some reason, it's failing to identify special cases in the MSVC 2015
build on AppVeyor. (Works on gcc and MSVC 2022...)

TODO: It might actually be the std::array<> conversion. If this doesn't
fix it, I'll revert that next.

This partially reverts commit e0d964098c5ae9ebfab7060fdaba178bd3d493db.
(Converted a lot more C arrays to std::array<>.)