[gtk] rp-tracker.cpp: Add a shutdown function.
The shutdown function ensures the dlopen()'d module handles are dlclose()'d.

NOTE: We can't add an init function because it had an extra parameter
for thread awareness prior to 1.11.2. We can't tell if it's v1 or v2+
at this point, so don't bother.

Function pointers will be initialized when
tracker_extract_get_metadata() is called.