[tracker] tracker-file-utils.c: dlopen("libblkid.so.1")
Only two functions are needed from libblkid:
- blkid_get_cache()
- blkid_get_tag_value()

dlopen() the library instead of linking directly to it, and define the
necessary structs and function prototypes here. This removes the need
for the libblkid-dev dependency at build time.

NOTE: libblkid.so.1 is *not* dlclose()'d. This shouldn't be a problem,
since Tracker uses it directly anyway.