[libromdata] NintendoDS_BNR: Verify CRC16s.
This should help prevent false-positives, since the CRC16s must be
correct in order for the DS firmware to accept the icon/title.

CRC16 is a fast algorithm, so this shouldn't cause a significant
performance hit. Also, it's only checked if the version number
matches one of the known versions.