[librpbase] IDiscReader: Add std::enable_shared_from_this().
This adds an std::weak_ptr<> object to the class that's populated once
the first std::shared_ptr<> is created for the object.

This allows us to give PartitionFile objects an IDiscReaderPtr to an
IDiscReader instead of a raw IDiscReader*.

GdiReader::openIsoPartition(): The PartitionFilePtr isn't needed
due to the IDiscReader rewrite that makes it a subclass of IRpFile.