[libromdata] GameCubeSave: Improve icon delays.
The icon delays are 0, 4 frames, 8 frames and 12 frames. The previous
method of using 125ms as the fastest speed was incorrect; it should be
67ms NTSC, 80ms PAL.

Also, use 50 Hz timing if the game ID ends in 'P'. The frame count is
still 0, 4, 8, and 12, but divided by 50 instead of 60.

In addition, for some reason, GCN IPL and Wii System Menu add an extra
frame to the first image, and subtract a frame from the last image.
Hence, a 2-icon animation using the 12-frame speed (e.g. SA2B) actually
shows 13, 11.

TODO: Need to adjust the first icon length for "bounce" in some cases?

Thanks to @Louis322 for noticing this error.