[gtk3] Add Nemo's name and description provider interface.
This is used by Nemo's plugin manager to show a useful desrciption for
each plugin, i.e. the property page provider and the menu extension.

NOTE: The extra "RpNautilusPropertyPageProvider:::" is needed, even
though it results in that being duplicated in nemo-extensions-list.
Extensions in the nemo-extensions repository do this. I'm not sure
why nemo-extensions-list duplicates it, but if it isn't present, then
the useful description isn't displayed.