[librpbase] RomFields: Add templated strArrayToVector*() overloads that take an std::array<>.
Most uses of strArrayToVector() and strArrayToVector_i18n() now use an
std::array<> as the source. std::array<const char*, size>  are not
compatible with arrays with different sizes.

To work around this, the overload itself uses a templated value for the
size parameter. It then calls the original strArrayToVector() function
using .data() and .size().

The end result is we only need to specify the std::array<> once in the
function call, and the template will automatically get the .data() and
.size() parameters.

Effectively no code size changes; this just helps to make the original
source code more readable.