[libromdata] CdiReader: Read the indexes correctly and attempt to handle tracks with subchannels.
The indexes are stored in a variable-length array. If only one index is
present, assume 0 pregap and use it as the entire track length. If two
indexes are present, the first is pregap and the second is data.

The subchannel code is untested, but it "should" work, since the
subchannels are located after the main sector data.

The "sector read mode" value is now stored in BlockRange, but it isn't
actually used for anything right now.

TODO: Skip CD-Text blocks.