[librptexture] KhronosKTX2.cpp: New inline helper function read_le32().
On i386/amd64, this is essentially the same code, since reading unaligned
data is allowed and is usually fast.

On other CPUs, reading unaligned data is either not allowed or is slow.
Use byte reads and shifting instead.

Effectively no change on i386/amd64, but this may help fix potential
crashes on other CPUs.

On ARMv5 and earlier, unaligned accesses crash.
On ARMv6 and later, it probably won't crash, but it might be slow.