[librpbase] RomFields, RomMetaData: Add begin()/end() aliases that return const_iterator.
These are merely wrappers around cbegin()/cend().

Range-based `for` loops only support begin()/end(), so adding these
aliases allows us to mak euse of range-based `for` loops.

Update other classes that make use of RomFields and RomMetaData
to use range-based `for` loops.

TODO: GTK RomDataView.cpp can't be updated to use it yet because of
the way it checks for "last field on the tab". Need to rework it.