[minizip-ng] Build as a .dll on Windows.
Since we're now using MiniZip in the J2ME parser, it should be compiled
as a DLL instead of statically linked.

Disable zstd support. It's causing issues with DelayLoad, since we're not
loading zstd in the J2ME parser. I originally enabled this to potentially
improve compression for test cases, but I ended up using .tar.zst instead
of zstd-compressed .zip files.

[libwin32common] Add MiniZip DLL information.

[libromdata] J2ME: Add MiniZip delay-load handling.

[libromdata] GczReader: Improve delay-load handling a bit.